Criando app cliente e organizando models
