--- funcionarios/urls.py
@@ -1,4 +1,0 @@
-from django.urls import path
-from funcionarios.views import view_cadastrar_funcionario, view_cadastrar_veterinario
-
-urlpatterns = []
