+++ modelos.py
@@ -1,26 +1,4 @@
-
 from django.db import models
-
-
-class Animal(models.Model):
-    nome = models.CharField(max_length=50, blank=True, null=True)
-    sexo = models.CharField(max_length=1, blank=True, null=True)
-    especie = models.CharField(max_length=50)
-    raca = models.CharField(max_length=50)
-    cor = models.CharField(max_length=50, blank=True, null=True)
-    datanasc = models.DateField(blank=True, null=True)
-    observacao = models.CharField(max_length=200, blank=True, null=True)
-    microchip = models.CharField(unique=True, max_length=50, blank=True, null=True)
-    cpf_cliente = models.ForeignKey('Cliente', models.DO_NOTHING, db_column='cpf_cliente')
-
-    class Meta:
-        managed = False
-        db_table = 'ANIMAL'
-
-    def __str__(self):
-        return self.cpf_cliente + '-' + self.nome
-
-
 
 class Atendimento(models.Model):
     observacao = models.CharField(max_length=100, blank=True, null=True)
@@ -34,9 +12,6 @@ class Atendimento(models.Model):
 
     def __str__(self):
         return self.id
-
-
-
 
 class AtendimentoProcClinico(models.Model):
     id_atendimento = models.ForeignKey(Atendimento, models.DO_NOTHING, db_column='id_atendimento', primary_key=True)
@@ -65,9 +40,6 @@ class AtendimentoProcEstetico(models.Model):
     def __str__(self):
         return self.id_atendimento + ' ' + self.id_proc_estetico
 
-
-
-
 class Autorizacao(models.Model):
     link_doc = models.CharField(unique=True, max_length=500, blank=True, null=True)
     id_proc_clinico = models.ForeignKey('ProcedimentoClinico', models.DO_NOTHING, db_column='id_proc_clinico')
@@ -78,55 +50,6 @@ class Autorizacao(models.Model):
 
     def __str__(self):
         return self.id
-
-
-
-class Cargo(models.Model):
-    nome = models.CharField(max_length=50)
-
-    class Meta:
-        managed = False
-        db_table = 'CARGO'
-
-    def __str__(self):
-        return self.id
-
-
-
-class CargoFuncionario(models.Model):
-    id_cargo = models.ForeignKey(Cargo, models.DO_NOTHING, db_column='id_cargo', primary_key=True)
-    id_func = models.ForeignKey('Funcionario', models.DO_NOTHING, db_column='id_func')
-
-    class Meta:
-        managed = False
-        db_table = 'CARGO_FUNCIONARIO'
-        unique_together = (('id_cargo', 'id_func'),)
-
-    def __str__(self):
-        return self.id_cargo + ' ' + self.id_func
-
-
-
-
-class Cliente(models.Model):
-    cpf = models.CharField(primary_key=True, max_length=14)
-    nome = models.CharField(max_length=50)
-    email = models.CharField(max_length=50, blank=True, null=True)
-    logradouro = models.CharField(max_length=50)
-    bairro = models.CharField(max_length=50)
-    cidade = models.CharField(max_length=50)
-    cep = models.CharField(max_length=8, blank=True, null=True)
-    numero = models.SmallIntegerField()
-    complemento = models.CharField(max_length=100, blank=True, null=True)
-    id_tipo_cliente = models.ForeignKey('TipoCliente', models.DO_NOTHING, db_column='id_tipo_cliente')
-
-    class Meta:
-        managed = False
-        db_table = 'CLIENTE'
-
-    def __str__(self):
-        return self.cpf
-
 
 class Comissao(models.Model):
     valor = models.DecimalField(max_digits=10, decimal_places=2)
@@ -200,23 +123,6 @@ class FeitoPor(models.Model):
     def __str__(self):
         return self.id_atendimento + ' ' + self.id_funcionario
 
-
-
-class FichaAnimal(models.Model):
-    id_animal = models.ForeignKey(Animal, models.DO_NOTHING, db_column='id_animal', primary_key=True)
-    data_consulta = models.DateField()
-    descricao = models.CharField(max_length=500, blank=True, null=True)
-
-    class Meta:
-        managed = False
-        db_table = 'FICHA_ANIMAL'
-        unique_together = (('id_animal', 'data_consulta'),)
-
-    def __str__(self):
-        return self.id
-
-
-
 class FichaDiagnostico(models.Model):
     id_diagnostico = models.IntegerField(primary_key=True)
     data_consulta = models.ForeignKey(FichaAnimal, models.DO_NOTHING, db_column='data_consulta')
@@ -230,22 +136,6 @@ class FichaDiagnostico(models.Model):
     def __str__(self):
         if __name__ == '__main__':
             return self.id_animal + ' ' + self.id_diagnostico
-
-
-
-class Funcionario(models.Model):
-    id = models.ForeignKey('User', models.DO_NOTHING, primary_key=True)
-    cpf = models.CharField(unique=True, max_length=11)
-    data_nasc = models.DateField()
-    equipe_sistema = models.IntegerField()
-
-    class Meta:
-        managed = False
-        db_table = 'FUNCIONARIO'
-
-    def __str__(self):
-        return self.id
-
 
 class Orcamento(models.Model):
     preco_final = models.DecimalField(max_digits=10, decimal_places=2)
@@ -314,20 +204,6 @@ class Responsavel(models.Model):
     def __str__(self):
         return self.cpf
 
-
-class StatusAnimal(models.Model):
-    id_status = models.ForeignKey('TipoStatusAnimal', models.DO_NOTHING, db_column='id_status', primary_key=True)
-    id_animal = models.ForeignKey(Animal, models.DO_NOTHING, db_column='id_animal')
-
-    class Meta:
-        managed = False
-        db_table = 'STATUS_ANIMAL'
-        unique_together = (('id_status', 'id_animal'),)
-
-    def __str__(self):
-        return self.id_animal + ' ' + self.id_status
-
-
 class StatusAtendimento(models.Model):
     id_atendimento = models.ForeignKey(Atendimento, models.DO_NOTHING, db_column='id_atendimento', primary_key=True)
     id_status = models.ForeignKey('TipoStatusAtendimento', models.DO_NOTHING, db_column='id_status')
@@ -353,33 +229,6 @@ class StatusEstadia(models.Model):
 
     def __str__(self):
         return self.id_estadia + ' ' + self.id_status
-
-
-
-class TelefoneCliente(models.Model):
-    cpf_cliente = models.ForeignKey(Cliente, models.DO_NOTHING, db_column='cpf_cliente', primary_key=True)
-    telefone = models.CharField(max_length=11)
-
-    class Meta:
-        managed = False
-        db_table = 'TELEFONE_CLIENTE'
-        unique_together = (('cpf_cliente', 'telefone'),)
-
-    def __str__(self):
-        return self.cpf_cliente + ' ' + self.telefone
-
-
-
-class TipoCliente(models.Model):
-    nome = models.CharField(max_length=50, blank=True, null=True)
-
-    class Meta:
-        managed = False
-        db_table = 'TIPO_CLIENTE'
-
-    def __str__(self):
-        return self.id
-
 
 class TipoDiagnostico(models.Model):
     id = models.SmallIntegerField(primary_key=True)
@@ -416,18 +265,6 @@ class TipoProcedimento(models.Model):
     def __str__(self):
         return self.id
 
-
-class TipoStatusAnimal(models.Model):
-    nome = models.CharField(max_length=50, blank=True, null=True)
-
-    class Meta:
-        managed = False
-        db_table = 'TIPO_STATUS_ANIMAL'
-
-    def __str__(self):
-        return self.id
-
-
 class TipoStatusAtendimento(models.Model):
     nome = models.CharField(max_length=50, blank=True, null=True)
 
@@ -450,25 +287,6 @@ class TipoStatusEstadia(models.Model):
     def __str__(self):
         return self.id
 
-
-class User(models.Model):
-    password = models.CharField(max_length=128)
-    last_login = models.DateTimeField(blank=True, null=True)
-    primeiro_nome = models.CharField(max_length=50)
-    ultimo_nome = models.CharField(max_length=50)
-    login = models.CharField(unique=True, max_length=150)
-    is_staff = models.IntegerField()
-    is_superuser = models.IntegerField()
-    situacao = models.IntegerField()
-
-    class Meta:
-        managed = False
-        db_table = 'USER'
-
-    def __str__(self):
-        return self.login
-
-
 class UserGroups(models.Model):
     user = models.ForeignKey(User, models.DO_NOTHING)
     group = models.ForeignKey('AuthGroup', models.DO_NOTHING)
@@ -487,23 +305,6 @@ class UserUserPermissions(models.Model):
         managed = False
         db_table = 'USER_user_permissions'
         unique_together = (('user', 'permission'),)
-
-
-class Veterinario(models.Model):
-    cpf = models.CharField(unique=True, max_length=11)
-    data_nasc = models.DateField()
-    equipe_sistema = models.IntegerField()
-    primeiro_nome = models.CharField(max_length=50)
-    ultimo_nome = models.CharField(max_length=50)
-    crm = models.CharField(unique=True, max_length=50)
-    estado_emissor = models.CharField(max_length=2)
-
-    class Meta:
-        managed = False
-        db_table = 'VETERINARIO'
-
-    def __str__(self):
-        return self.cpf
 
 
 class AuthGroup(models.Model):
