+++ usuarios/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from usuarios.models import User
+from usuarios.models import User, Funcionario, Veterinario
 
 
 class FormUser(forms.ModelForm):
@@ -22,3 +22,52 @@ class FormAlteraUser(forms.ModelForm):
     class Meta:
         model = User
         fields = ['primeiro_nome', 'ultimo_nome', 'password']
+
+
+class FormFuncionario(forms.ModelForm):
+
+    def save(self, commit=True):
+        usuario = super(FormFuncionario, self).save(commit=False)
+        usuario.set_password("123@mudar")
+        if commit:
+            usuario.save()
+        return usuario
+
+    class Meta:
+        model = Funcionario
+        fields = [
+            'primeiro_nome',
+            'ultimo_nome',
+            'cpf',
+            'data_nasc',
+            'equipe_sistema'
+        ]
+
+
+class FormAlteraFuncionario(forms.ModelForm):
+
+    class Meta:
+        model = Funcionario
+        fields = [
+            'primeiro_nome',
+            'ultimo_nome',
+            'cpf',
+            'data_nasc',
+            'equipe_sistema',
+            'password'
+        ]
+
+
+class FormVeterinario(forms.ModelForm):
+
+    class Meta:
+        model = Veterinario
+        fields = [
+            'primeiro_nome',
+            'ultimo_nome',
+            'cpf',
+            'data_nasc',
+            'equipe_sistema',
+            'crm',
+            'estado_emissor'
+        ]
